Saving reading time in milliseconds

+++ src/main/java/com/example/readerai/dto/statistics/ReadingStatsByDay.java
@@ -22,7 +22,7 @@ public class ReadingStatsByDay {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class DailyReadingStats {
-        private Long totalReadingTimeMinutes;
+        private float totalReadingTimeMinutes;
         private Integer totalPagesRead;
     }
 }
